Goodbye, glium.
I'm really disappointed, because this API looked so promising at first, but
it's actually quite flawed.

Most of this seems to arise from the library trying to do too many things.

- Make a totally memory-safe abstraction. This is a very rustic thing to do and
  is probably what we'll be doing later anyway. It's also not that difficult
  for the most part.

- Provide RAII for GL handles. Also a really good thing and not too difficult.

- Provide total safety. This is impossible because of shaders, short of
  defining a new, weaker shader language. (Though there's some hints in the
  docs that the devs might be thinking just that...)

- Provide _error_ safety. That is, glium tries to make it impossible for
  `glGetError()` to ever return an error, instead doing its own error checking.
  While some of this is achieved by virtue of having a nicer interface, it also
  results in some contortions.

- Abstract away the state machine. A nominally noble goal, but it means we end
  up emulating a stateless system on top of an emulation of a state machine.

- Provide an automatic and type-safe way to declare vertices and uniforms. I
  list this last not because it's a bad idea, but it's where the library
  completely fails.

The most egregious issue is that of uniforms. This involves using that
`uniform!` macro at every shader call site. This results in an abstraction
that's *less type-safe than C*, because there will be no compile-time error if
one site happens to misspell a field name. This problem is made worse by the
fact that glium tolerates providing fields that don't bind to anything in the
shader.

Digging into the source also reveals that uniforms are bound _by name_, and are
bound afresh _every time a drawing operation occurs_.

While vertices have a reasonable nterface, perhaps by virtue of OpenGL forcing
them to be offset-based structs, there's still another huge failing: There is
nothing preventing anyone from passing an incompatible vertex type to a shader
at compile time. Again, this is _less type-safe than C_.

The fundamental failing here is that a shader inherently defines the uniforms
and vertices it takes, but glium makes no attempt to make that connection (and
actuall divorces uniforms further than native GL).

While I could see the glium dev(s) being receptive to fixing the type-safety
issues, there's some clear philosophical disparity here, so it's honestly not
worth it to try to fork and fix that issue.

New plan is to use GLEW and port Mantigraphia's abstraction layer to Rust.