Try to make things run on GL ES 1.0.
I presume the reason it doesn't work is due to trying to use VAOs, but
it also returns errors for some of the texture things.

I was really hoping to not need to delve into the hell that is an
OpenGL-2.0-style API. It's probably not worth supporting ES for now.